support additional unpacking generalizations; starargs, kwargs departure
The changes codified in PEP 448 for Python 3.5 allow the
iterable- ("splat") and dictionary- ("double splat") unpacking arguments
to be used an arbitrary number of times in calls. As a result of this,
the starargs and kwargs attributes on some ast nodes have vanished,
their function superceded by ast.Starred objects in inside of `args` and
ast.keyword objects with a name of None inside of `keywords`,
respectively.