non-hideous dictionary displays
The trailing comma-space that was previously output, while perfectly
legal in every way, is not what most human Python programmers would
write in most situations. We use the same slick idiom as the
`comma_list` method (iterating over the enumeration, but writing the
comma-space before the elements if the index is truthy, which zero is
not) to not put anything extra after the final key-value pair.