Yield fixes
It turns out the cheezy way I did yield fixes before was no good, because
it is legal to do (yield foo)(bar).

The new way always encloses yield statements in parentheses.  This is ugly,
and MAY not (I don't know) work on earlier versions, although it works fine
on 2.7.