Fix trailing newlines in f-strings
If an f-string was ended with a newline character (more precisely, if
the last element of a JoinedStr node was just a newline character), its
round-trip source code representation would have an extra newline on a
wrong place, resulting in invalid syntax.

The bug affects only python>=3.6 (f-strings weren't supported before)

This commit fixes the aforementioned bug.

Fixes #89.