Extract ASTs from methods and nested functions
Previously ASTs could be extracted only for bare functions. By
recursively searching "blocks" for function definitions, methods and
nested functions can be located. In this context a "block" is generally
an `ast.AST` subclass with a `body` field, such as `ast.ClassDef` and
`ast.FunctionDef`. Several other nodes such as `ast.If`, `ast.For`, etc
are also searched. See `CodeToAst.block_types` for the complete list.